+++ udf_ui_categories_view.py
@@ -5,8 +5,15 @@ from PyQt6.QtCore import QRect, pyqtSignal
 from PyQt6.QtWidgets import QCheckBox, QDialog, QGridLayout, QPushButton, QWidget
 
 
-class UbuntuDesktopFileCategories(QDialog):
-    # Signal emitted when categories are selected
+class UbuntuDesktopFileCategoriesView(QDialog):
+    '''Manage the Ubuntu Desktop File Categories View.
+
+    This class represents the dialog window for selecting categories. 
+    It provides checkboxes for predefined categories and emits a signal with the selected categories.
+
+    Attributes:
+        categories_selected: A signal emitted when categories are selected.'''
+
     categories_selected = pyqtSignal(list)
 
     # Predefined categories
@@ -52,18 +59,18 @@ class UbuntuDesktopFileCategories(QDialog):
         self.pushButton = QPushButton(self)
         self.pushButton.setText("Ok")
         self.pushButton.setGeometry(QRect(int((self.width() / 2) - 34), 168, 68, 32))
-        self.pushButton.clicked.connect(self.retrieve_selected_categories)
+        self.pushButton.clicked.connect(self.close)
 
-    def retrieve_selected_categories(self) -> None:
+    def _get_type_categories(self) -> list[str]:
         # Retrieve the selected categories from the checkboxes
-        list_categories = [
+        return [
             check_box.text()
             for check_box in self.gridLayoutWidget.findChildren(QCheckBox)
             if check_box.isChecked()
         ]
 
+    def close(self) -> None:
+        list_categories = self._get_type_categories()
         # Emit the selected categories as a signal
         self.categories_selected.emit(list_categories)
-
-        # Close the dialog window
-        self.close()
+        super().close()
